requirements.txt
click==7.1.2
PyYAML==5.4.1
jinja2==2.11.3
